feat: format negative numbers correctly closes #178

+++ internal/common/ui.go
@@ -32,12 +32,12 @@ const (
 
 // SI prefixes
 const (
-	K int64 = 1e3
-	M int64 = 1e6
-	G int64 = 1e9
-	T int64 = 1e12
-	P int64 = 1e15
-	E int64 = 1e18
+	K float64 = 1e3
+	M float64 = 1e6
+	G float64 = 1e9
+	T float64 = 1e12
+	P float64 = 1e15
+	E float64 = 1e18
 )
 
 // FormatNumber returns number as a string with thousands separator
